refactor : 컴포넌트 파일 이름 변경
